app and bundle name, splash screen, and build generated
