+++ android/app/build.gradle
@@ -74,15 +74,23 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.mymeals"
+    namespace "com.mingossi.mymeals"
     defaultConfig {
-        applicationId "com.mymeals"
+        applicationId "com.mingossi.mymeals"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
         versionName "1.0"
     }
     signingConfigs {
+        release {
+           if(project.hasProperty('MYAPP_UPLOAD_STORE_FILE')) {
+              storeFile file(MYAPP_UPLOAD_STORE_FILE)
+              storePassword MYAPP_UPLOAD_STORE_PASSWORD
+              keyAlias MYAPP_UPLOAD_KEY_ALIAS
+              keyPassword MYAPP_UPLOAD_KEY_PASSWORD
+           }
+        }
         debug {
             storeFile file('debug.keystore')
             storePassword 'android'
@@ -97,6 +105,7 @@ android {
         release {
             // Caution! In production, you need to generate your own keystore file.
             // see https://reactnative.dev/docs/signed-apk-android.
+            // CHANGE HERE TO RELEASE BEFORE GENERATING NEW BUILD
             signingConfig signingConfigs.debug
             minifyEnabled enableProguardInReleaseBuilds
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
