+++ android/app/src/main/java/com/mingossi/mymeals/MainActivity.kt
@@ -1,9 +1,11 @@
-package com.mymeals
+package com.mingossi.mymeals
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
+import android.os.Bundle;
+import com.zoontek.rnbootsplash.RNBootSplash;
 
 class MainActivity : ReactActivity() {
 
@@ -19,4 +21,9 @@ class MainActivity : ReactActivity() {
    */
   override fun createReactActivityDelegate(): ReactActivityDelegate =
       DefaultReactActivityDelegate(this, mainComponentName, fabricEnabled)
+
+  override fun onCreate(savedInstanceState: Bundle?) {
+      RNBootSplash.init(this, R.style.BootTheme) // ⬅️ initialize the splash screen
+      super.onCreate(savedInstanceState) // super.onCreate(null) with react-native-screens
+    }
 }
