+++ assets/bootsplash/manifest.json
@@ -1,0 +1,7 @@
+{
+  "background": "#3fae4d",
+  "logo": {
+    "width": 200,
+    "height": 190
+  }
+}
