+++ ios/MyMeals/Colors.xcassets/BootSplashBackground-48333a.colorset/Contents.json
@@ -1,0 +1,20 @@
+{
+  "colors": [
+    {
+      "idiom": "universal",
+      "color": {
+        "color-space": "srgb",
+        "components": {
+          "blue": "0.301960784313725",
+          "green": "0.682352941176471",
+          "red": "0.247058823529412",
+          "alpha": "1.000"
+        }
+      }
+    }
+  ],
+  "info": {
+    "author": "xcode",
+    "version": 1
+  }
+}
