+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './handleTotals.ts';
 export * from './calendarUtils.ts';
 export * from './formatNumberWithComma.ts';
 export * from './allMealsTotals.ts';
+export * from './profileUtils.ts';
